Add funtion was returning some format error, but it's fixed now. Then I added an exit functionality. So users can exit the Add() anytime by entering -- (double minus/hyphen). Will work on the DateTime next.
I also discovered that ReSharper is indicating possible mutiple enumeration of IEnumerable in some of the functions we have in Linq.cs, if I remember very well, the problem should be solved by returning a list from the LINQ query.
Although ALL the functions works perfectly, it will be nice to fix that subtle "bug".

I'll try and look into it if I can find a solution. Hopefully!